pyproject.toml
[project]
name = "my-blogger-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.6.3",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.13",
    "lxml-html-clean>=0.4.2",
    "newspaper4k>=0.9.3.1",
    "openai>=1.90.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
